AOE Damage

+++ app/game/entities/arrow.cpp
@@ -1,5 +1,7 @@
 #include "arrow.h"
 #include "game/entities/player.h"
+#include "game/entities/gladiator.h"
+#include "game/grid.h"
 
 Arrow::Arrow(QWidget *parent, int x, int y) : QFrame(parent)
 {
@@ -20,7 +22,7 @@ Arrow::Arrow(QWidget *parent, int x, int y) : QFrame(parent)
     setStyleSheet("background-color:#635255;"
                   "image: url(:/img/fire/00fire.png)");
 
-//    this->setStyleSheet("background-color:#EBF5EE;");
+    //    this->setStyleSheet("background-color:#EBF5EE;");
     this->setGeometry(x, y, width, height);
     this->show();
 }
@@ -43,15 +45,53 @@ void Arrow::draw()
 
 void Arrow::collide()
 {
+    tower->setType(3);
     for (Entity *entity : GameController::getInstance()->getEntities())
     {
         if (entity->tag == "player" &&
                 Collision::collide(getRect(), entity->getRect()))
         {
             Player *player = dynamic_cast<Player *>(entity);
-            tower->setXp(tower->getXp() + 1);
-            player->kill();
-            kill();
+            int damageDone = tower->getDamagePerShoot() - player->getGladiator()->getThoughness();
+            if(damageDone < 1){
+                damageDone = 1;
+            }
+            if(tower->getType() != 3){
+                player->getGladiator()->setHealth(player->getGladiator()->getHealth() - damageDone);
+                if(player->getGladiator()->getHealth() <= 0){
+                    tower->setXp(tower->getXp() + 1);
+                    player->kill();
+                    kill();
+                }
+            }
+            else{
+                QList<Entity *> playersHit = AOECollide();
+                QList<Entity *>::iterator i;
+                for(i = playersHit.begin(); i != playersHit.end(); i++){
+                    Player *tempPlayer = dynamic_cast<Player *>(*i);
+                    tempPlayer->getGladiator()->setHealth(player->getGladiator()->getHealth() - damageDone);
+                    if(tempPlayer->getGladiator()->getHealth() <= 0){
+                        tower->setXp(tower->getXp() + 1);
+                        tempPlayer->kill();
+
+                        cout << "h" << endl;
+                    }
+                }
+                playersHit.clear();
+                cout << "hh" << endl;
+                kill();
+
+                Grid *grid = dynamic_cast<Grid *>(parent());
+                QFrame *qFrame = new QFrame(grid);
+                qFrame->setStyleSheet("background-color:#635255;");
+                int xPoss = x - grid->getTileSize();
+                int yPoss = y - grid->getTileSize();
+                int diameter = grid->getTileSize() * 3;
+                qFrame->setGeometry(xPoss,yPoss,diameter,diameter);
+                qFrame->show();
+
+                QTimer::singleShot(1000, qFrame, &QFrame::close);
+            }
             break;
         }
     }
@@ -130,4 +170,27 @@ void Arrow::move()
     y = Math::approach(y, targetPlayer->getY() + offset, ySpeed);;
 
     QFrame::move(x, y);
+}
+
+QRegion Arrow::getCircle()
+{
+    Grid *grid = dynamic_cast<Grid *>(parent());
+    int xPoss = x - grid->getTileSize();
+    int yPoss = y - grid->getTileSize();
+    int diameter = grid->getTileSize() * 3;
+    QRegion circle(xPoss, yPoss, diameter, diameter, QRegion::Ellipse);
+    return circle;
+}
+
+QList<Entity *> Arrow::AOECollide(){
+    QList<Entity *> players;
+    for (Entity *entity : GameController::getInstance()->getEntities())
+    {
+        if (entity->tag == "player" &&
+                Collision::collide(getCircle(), entity->getRect()))
+        {
+            players.push_back(entity);
+        }
+    }
+    return players;
 }
